# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SprintPulse API
  version: 0.2.0
  license:
    name: Private hackathon prototype
  description: |
    OpenAPI contract for the Express routes mounted at `/api` in SprintPulse.

    The API currently uses a `personaId` query or body field to resolve the acting user. Some routes are available only
    when `ENABLE_MOCK_FLOW=true`; in Supabase data mode those routes return `501` until their adapters are implemented.
servers:
  - url: http://localhost:4000/api
    description: Local SprintPulse API
...
security: []
paths:
...
  /projects/{projectId}/standups:
    get:
      tags: [Standups]
      operationId: listProjectStandups
      summary: List standups for a project sprint.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
        - $ref: "#/components/parameters/PersonaIdQuery"
        - $ref: "#/components/parameters/SprintIdQuery"
      responses:
        "200":
          description: Project standups. Mock flow returns a smaller project-plus-standups payload.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ProjectStandupsResponse"
                  - $ref: "#/components/schemas/MockProjectStandupsResponse"
        "404":
          description: Standups are not found or not visible.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notVisible:
                  value:
                    error: Standups not found or not visible to this user
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: [Standups]
      operationId: submitProjectStandup
      summary: Submit a manual standup for a project.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
      requestBody:
        $ref: "#/components/requestBodies/StandupSubmitBody"
      responses:
        "201":
          description: Standup accepted and member pulse refreshed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStandupSubmitResponse"
        "400":
          $ref: "#/components/responses/StandupValidation"
        "403":
          description: Persona cannot submit standups for this project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                forbidden:
                  value:
                    error: You do not have permission to submit standups for this project
        "500":
          $ref: "#/components/responses/InternalServerError"
  /projects/{projectId}/transcripts/parse:
    post:
      tags: [Standups]
      operationId: parseProjectTranscript
      summary: Parse transcript text into project standup entries.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectTranscriptParseRequest"
            examples:
              transcript:
                value:
                  personaId: priya
                  transcript: "Priya: Yesterday I finalized sprint goals. Today I am reviewing risks. Blockers none."
      responses:
        "200":
          description: Transcript parsed into standup entries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectTranscriptParseResponse"
        "400":
          $ref: "#/components/responses/TranscriptValidation"
        "404":
          description: Project was not found in mock flow.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                projectNotFound:
                  value:
                    error: Project not found
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /projects/{projectId}/transcripts/teams-webhook:
    post:
      tags: [Standups, Integrations]
      operationId: ingestTeamsTranscript
      summary: Ingest a Microsoft Teams meeting transcript and auto-create standups.
      description: |
        Webhook endpoint designed to be called by an external delivery
        mechanism (Power Automate flow, Otter.ai / Fireflies.ai bot, a custom
        Microsoft Graph poller, or any HTTP client) when a Teams meeting
        transcript becomes available. Accepts either WebVTT (the default Teams
        export format) or plain `Speaker:` text and feeds it through the same
        parser used by the manual paste flow.

        Authorization: this route is unauthenticated, but the `organizerEmail`
        must match a profile that is a member of the project. Requests from
        organizers outside the project return 404. For production hardening,
        gate this behind a shared-secret header (`X-SprintPulse-Webhook-Token`).

        Returns 501 when `ENABLE_MOCK_FLOW=true` since the route writes to
        Supabase.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamsWebhookRequest"
            examples:
              vtt:
                summary: Microsoft Teams WebVTT export
                value:
                  organizerEmail: yash@sprintpulse.dev
                  meetingSubject: SCRUM Daily Standup
                  meetingId: 19:meeting_abc123@thread.v2
                  transcript: |
                    WEBVTT

                    1
                    00:00:01.000 --> 00:00:08.000
                    <v Yash>Yesterday I shipped the dashboard refactor. Today I am wiring Teams ingestion. No blockers.

                    2
                    00:00:09.000 --> 00:00:14.000
                    <v Yanshi>Yesterday I reviewed AI scoring tests. Today I will polish the integrations page. Blocked on Jira keys.
              plainText:
                summary: Plain text with Speaker markers
                value:
                  organizerEmail: yash@sprintpulse.dev
                  meetingSubject: SCRUM Daily Standup
                  transcript: |
                    Yash: Yesterday I shipped the dashboard refactor. Today I am wiring Teams ingestion. No blockers.
                    Yanshi: Yesterday I reviewed AI scoring tests. Today I will polish the integrations page. Blocked on Jira keys.
      responses:
        "201":
          description: Transcript parsed and standup entries persisted to the active sprint.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamsWebhookResponse"
        "400":
          description: Required fields are missing from the body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                noTranscript:
                  value:
                    error: transcript is required
                noOrganizer:
                  value:
                    error: organizerEmail is required
        "404":
          description: Organizer is not a member of the target project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                organizerNotMember:
                  value:
                    error: "No project member matches organizer alice@example.com. The meeting organizer must be a member of this project."
        "500":
          $ref: "#/components/responses/InternalServerError"
        "501":
          description: Webhook is disabled because the API is running in mock-data mode.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                mockMode:
                  value:
                    error: Teams webhook is only available in real-data mode.
  /projects/{projectId}/standups/sync:
    post:
      tags: [Standups, Integrations]
      operationId: syncProjectStandups
      summary: Sync or backfill project standup entries.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
      requestBody:
        $ref: "#/components/requestBodies/PersonaActionBody"
      responses:
        "200":
          description: Standup sync completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandupsSyncResponse"
        "403":
          description: Persona cannot sync standups for this project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                forbidden:
                  value:
                    error: You do not have permission to sync standups for this project
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /transcripts/parse:
    post:
      tags: [Legacy Mock]
      operationId: parseLegacyTranscript
      summary: Parse transcript text into seeded standup drafts.
      description: Mock-flow route. In Supabase data mode this returns `501`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscriptParseRequest"
            examples:
              transcript:
                value:
                  transcript: "Priya: Yesterday finalized sprint goals. Today reviewing delivery risk. Blockers none."
      responses:
        "200":
          description: Parsed seeded transcript entries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptParseResponse"
        "400":
          $ref: "#/components/responses/TranscriptValidation"
        "501":
          $ref: "#/components/responses/NotImplemented"
...
components:
  parameters:
    PersonaIdQuery:
      name: personaId
      in: query
      required: true
      description: Persona/profile id used by SprintPulse to resolve viewer permissions.
      schema:
        type: string
      example: priya
    SprintIdQuery:
      name: sprintId
      in: query
      required: false
      description: Optional sprint id filter for Supabase-backed sprint data.
      schema:
        type: string
      example: sprint-semi-2026
    ProjectIdPath:
      name: projectId
      in: path
      required: true
      description: Project id or, in mock flow, case-insensitive project key.
      schema:
        type: string
      example: sp-core
...
  requestBodies:
...
    PersonaActionBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PersonaActionRequest"
          examples:
            actor:
              value:
                personaId: priya
    StandupSubmitBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/StandupSubmitRequest"
          examples:
            standup:
              value:
                personaId: yash
                yesterday: Finalized API contracts.
                today: Connecting scoring with dashboard data.
                blockers: No blocker.
  responses:
    InternalServerError:
      description: The route failed while loading or mutating data.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internal:
              value:
                error: Unable to load data
...
    NotImplemented:
      description: Route is not available in the current runtime mode.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notImplemented:
              value:
                error: Supabase data mode is enabled, but this route still needs its database-backed adapter.
...
    StandupValidation:
      description: Standup body is missing required text fields.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missingStandupText:
              value:
                error: Yesterday and today fields are required
    TranscriptValidation:
      description: Transcript text is empty.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            transcriptRequired:
              value:
                error: Transcript text is required
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Persona not found
...
    HackathonRole:
      type: string
      enum: [frontend, backend, architect, qa]
...
    AccessScope:
      type: string
      enum: [team, individual, quality, presentation]
...
    Persona:
      type: object
      required: [id, name, email, initials, hackathonRole, productPersona, title, accessScope, focus]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        initials:
          type: string
        hackathonRole:
          $ref: "#/components/schemas/HackathonRole"
        productPersona:
          $ref: "#/components/schemas/ProductPersona"
        title:
          type: string
        accessScope:
          $ref: "#/components/schemas/AccessScope"
        focus:
          type: string
...
    SprintInfo:
      type: object
      required: [id, name, goal, startDate, endDate, status]
      properties:
        id:
          type: string
        name:
          type: string
        goal:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
          enum: [planned, active, closed]
    SprintProject:
      type: object
      required: [id, key, name, source, sprint, members, ownerIds, scrumMasterIds, createdBy, createdAt, updatedAt]
      properties:
        id:
          type: string
        key:
          type: string
        name:
          type: string
        source:
          $ref: "#/components/schemas/ProjectSource"
        jiraSite:
          type: string
        sprint:
          $ref: "#/components/schemas/SprintInfo"
        members:
          type: array
          items:
            $ref: "#/components/schemas/ProjectMember"
        ownerIds:
          type: array
          items:
            type: string
        scrumMasterIds:
          type: array
          items:
            type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastSyncAt:
          type: string
          format: date-time
...
    StandupEntry:
      type: object
      required: [id, memberId, date, yesterday, today, blockers, source]
      properties:
        id:
          type: string
        projectId:
          type: string
        memberId:
          type: string
        date:
          type: string
          format: date
        yesterday:
          type: string
        today:
          type: string
        blockers:
          type: string
        source:
          type: string
          enum: [manual, transcript, audio]
    StandupWithMember:
      allOf:
        - $ref: "#/components/schemas/StandupEntry"
        - type: object
          required: [memberName, memberInitials]
          properties:
            sprintId:
              type: string
            memberName:
              type: string
            memberInitials:
              type: string
...
    ProjectStandupsResponse:
      type: object
      required: [viewer, project, sprint, standups, canSubmit, canSync]
      properties:
        viewer:
          $ref: "#/components/schemas/Persona"
        project:
          $ref: "#/components/schemas/SprintProject"
        sprint:
          $ref: "#/components/schemas/SprintInfo"
        standups:
          type: array
          items:
            $ref: "#/components/schemas/StandupWithMember"
        canSubmit:
          type: boolean
        canSync:
          type: boolean
    MockProjectStandupsResponse:
      type: object
      required: [project, standups]
      properties:
        project:
          $ref: "#/components/schemas/SprintProject"
        standups:
          type: array
          items:
            $ref: "#/components/schemas/StandupEntry"
        aiScore:
          $ref: "#/components/schemas/AiMemberScore"
...
    PersonaActionRequest:
      type: object
      required: [personaId]
      properties:
        personaId:
          type: string
    StandupSubmitRequest:
      type: object
      required: [personaId, yesterday, today]
      properties:
        personaId:
          type: string
        yesterday:
          type: string
        today:
          type: string
        blockers:
          type: string
          default: No blocker.
...
    ProjectStandupSubmitResponse:
      type: object
      required: [entry, member, project]
      properties:
        entry:
          oneOf:
            - $ref: "#/components/schemas/StandupEntry"
            - $ref: "#/components/schemas/StandupWithMember"
        member:
          $ref: "#/components/schemas/MemberPulse"
        project:
          $ref: "#/components/schemas/SprintProject"
    TranscriptParseRequest:
      type: object
      required: [transcript]
      properties:
        transcript:
          type: string
    ProjectTranscriptParseRequest:
      allOf:
        - $ref: "#/components/schemas/TranscriptParseRequest"
        - type: object
          properties:
            personaId:
              type: string
...
    TranscriptParseResponse:
      type: object
      required: [mode, note, parsed]
      properties:
        mode:
          type: string
          example: transcript-parser
        note:
          type: string
        parsed:
          type: array
          items:
            $ref: "#/components/schemas/ParsedTranscriptEntry"
        analysis:
          $ref: "#/components/schemas/DailyStatusAnalysis"
        aiMeta:
          $ref: "#/components/schemas/AiGenerationMeta"
    ProjectTranscriptParseResponse:
      allOf:
        - $ref: "#/components/schemas/TranscriptParseResponse"
        - type: object
          required: [project]
          properties:
            project:
              $ref: "#/components/schemas/SprintProject"
...
    TeamsWebhookRequest:
      type: object
      required: [organizerEmail, transcript]
      properties:
        organizerEmail:
          type: string
          format: email
          description: |
            UPN / email of the meeting organizer. Must match the `email`
            field on an active profile that belongs to the target project.
            Acts as the implicit authorization signal — requests from
            organizers outside the project are rejected with 404.
          example: yash@sprintpulse.dev
        transcript:
          type: string
          description: |
            Transcript body. Accepts WebVTT (with `WEBVTT` header, cue
            numbers, timestamp ranges, and `<v Speaker>` tags) or plain
            text with `Speaker:` markers. Both formats flow through the
            same server-side parser.
        meetingSubject:
          type: string
          description: Title of the Teams meeting (optional, echoed back in response).
          example: SCRUM Daily Standup
        meetingId:
          type: string
          description: |
            Stable Teams meeting identifier (optional). When provided, it
            can be used by callers to deduplicate retries.
          example: 19:meeting_abc123@thread.v2
    TeamsWebhookResponse:
      allOf:
        - $ref: "#/components/schemas/ProjectTranscriptParseResponse"
        - type: object
          required: [source, organizerEmail]
          properties:
            source:
              type: string
              enum: [teams-webhook]
              description: Constant value identifying this ingestion path.
            organizerEmail:
              type: string
              format: email
            meetingSubject:
              type: string
            meetingId:
              type: string
    StandupsSyncResponse:
      type: object
      required: [project, syncedAt, importedStandups, warnings]
      properties:
        project:
          $ref: "#/components/schemas/SprintProject"
        syncedAt:
          type: string
          format: date-time
        importedStandups:
          type: integer
          minimum: 0
        warnings:
          type: array
          items:
            type: string
        linkedSyncs:
          type: array
          description: Parallel evidence sync results for Jira and Git when standups are synced.
          items:
            $ref: "#/components/schemas/ProjectSignalSyncStatus"
...